WIP switched to papaya - it's so slow I can't even test it

+++ src/lexer.rs
@@ -323,11 +323,7 @@ pub fn lex(content: &str, fi: Fi) -> Vec<SourceToken<'_>> {
         let line_after = line + content[scanned_to..s.end].matches('\n').count();
         scanned_to = s.end;
 
-        let span = Span::Known(
-            fi,
-            (line as usize, (s.start - indent) as usize),
-            (line_after as usize, (s.end - indent) as usize),
-        );
+        let span = Span::Known(fi, (line, s.start - indent), (line_after, s.end - indent));
         match t {
             Ok(Token::Indent(at)) => {
                 // We need to know the indentation of every token - even if there are tokens before it.
@@ -362,7 +358,7 @@ pub fn lex(content: &str, fi: Fi) -> Vec<SourceToken<'_>> {
         }
         line = line_after;
     }
-    let last_span = Span::Known(fi, (line as usize, 0), (line as usize, 0));
+    let last_span = Span::Known(fi, (line, 0), (line, 0));
     for (s, d) in state.iter() {
         if d.is_indented() && *s != (0, 0) {
             out.push((Ok(Token::LayEnd), last_span));
